fixed median calculation

+++ Main/MedianCalculator.hpp
@@ -4,5 +4,5 @@ namespace MedianCalculator
 {
     constexpr int SAMPLE_SIZE = 101;
     
-    void Update(const int newShortValue, const int newLongValue);
+    void Update(const double newShortValue, const double newLongValue);
 }
